move get_plane_grom_points to util
